+++ src/main/java/com/example/gojipserver/global/config/SecurityConfig.java
@@ -1,6 +1,6 @@
 package com.example.gojipserver.global.config;
 
-import com.example.gojipserver.domain.oauth2.service.CustomOAuth2UserService;
+import com.example.gojipserver.global.oauth2.service.CustomOAuth2UserService;
 import com.example.gojipserver.domain.oauth2.service.CustomUserDetailsService;
 import com.example.gojipserver.global.config.security.handler.OAuth2AuthenticationSuccessHandler;
 import com.example.gojipserver.global.config.security.jwt.JwtAuthenticationFilter;
@@ -31,20 +31,20 @@ public class SecurityConfig {
                 .httpBasic(httpBasic -> httpBasic.disable())
                 .rememberMe(rememberMe -> rememberMe.disable())
                 .sessionManagement(sessionManagement -> sessionManagement.sessionCreationPolicy(SessionCreationPolicy.STATELESS) // 세션을 사용하지 않기 때문에 세션 설정을 STATELESS로 설정
-        );
+                );
 
         http.authorizeRequests(authorizeRequests -> authorizeRequests
-                .requestMatchers("/oauth2/**").permitAll() // oauth2로 시작하는 모든 요청은 인증 없이 접근 가능
-                .requestMatchers("/login").permitAll() // login으로 시작하는 모든 요청은 인증 없이 접근 가능
-                .requestMatchers("/swagger-ui/index.html").permitAll()
+                        .requestMatchers("/oauth2/**").permitAll() // oauth2로 시작하는 모든 요청은 인증 없이 접근 가능
+                        .requestMatchers("/login").permitAll() // login으로 시작하는 모든 요청은 인증 없이 접근 가능
+                        .requestMatchers("/swagger-ui/index.html").permitAll()
 //                .anyRequest().authenticated() // 나머지 요청은 모두 인증 필요
         );
 
         http.oauth2Login(oauth->oauth
-                .authorizationEndpoint(authorization->authorization.baseUri("/oauth2/authorize")) // 인증 요청을 처리하는 Endpoint 설정
-                .redirectionEndpoint(redirection->redirection.baseUri("/oauth2/callback/*")) // OAuth2 인증 서버로부터 리다이렉트를 받을 Endpoint 설정
-                .userInfoEndpoint(userInfo->userInfo.userService(customOAuth2Service))
-                .successHandler(oAuth2AuthenticationSuccessHandler) //성공시
+                        .authorizationEndpoint(authorization->authorization.baseUri("/oauth2/authorize")) // 인증 요청을 처리하는 Endpoint 설정
+                        .redirectionEndpoint(redirection->redirection.baseUri("/oauth2/callback/*")) // OAuth2 인증 서버로부터 리다이렉트를 받을 Endpoint 설정
+                        .userInfoEndpoint(userInfo->userInfo.userService(customOAuth2Service))
+                        .successHandler(oAuth2AuthenticationSuccessHandler) //성공시
 //                .failureHandler(null) //실패시
         );
         http.userDetailsService(customUserDetailsService);
